+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## Unreleased
+
+* **Breaking change:** Previously, flake8-pyi monkey patched flake8's F821
+  (undefined name) check to avoid false positives in stub files. This monkey
+  patch has been removed, and we recommend to disable F821 when running flake8
+  on stub files.
+* Remove the now unnecessary `--no-pyi-aware-file-checker` option.
+
 ## 5.5.0
 
 New error codes: